requirements.txt
rembg
numpy
pillow
fastapi[standard]